server/app/modules/auth/schemas.py
from pydantic import BaseModel, Field


class LoginRequest(BaseModel):
    email: str = Field(min_length=3)
    password: str = Field(min_length=1)


class RefreshRequest(BaseModel):
    refresh_token: str = Field(min_length=1)


class LogoutRequest(BaseModel):
    refresh_token: str | None = None


class AuthUserPublic(BaseModel):
    id: str
    email: str
    full_name: str
    role: str


class TokenPairResponse(BaseModel):
    access_token: str
    refresh_token: str
    token_type: str = "bearer"
    user: AuthUserPublic


class LogoutResponse(BaseModel):
    revoked: bool
    session_id: str | None = None
