server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantflow-pro-server"
version = "0.1.0"
description = "FastAPI backend skeleton for QuantFlow Pro"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.5",
  "email-validator>=2.3.0",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "psycopg[binary]>=3.2.10",
  "pydantic-settings>=2.11.0",
  "pytest>=8.4.2",
  "redis>=6.4.0",
  "rq>=2.6.0",
  "sqlalchemy>=2.0.43",
  "uvicorn[standard]>=0.35.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
